requirements.txt
albumentations
Pillow
torch
tqdm
torchvision
numpy
tensorboard
rasterio